Change Icarus listener paradigm.
Previously, the flame model was
basically almost entirely mutable and
this made updating UI a bitch and I basically
couldn't have multiple windows open at once.

Updated things so that the model is contained within
a wrapper class that notifies dialogs upon a change in model
which means now each frame is responsible for knowing
how to update its content when there is a new model.